Made constants and handler packages internal in the lametrictime binding
