+++ addons/binding/org.openhab.binding.lametrictime/src/main/java/org/openhab/binding/lametrictime/internal/discovery/LaMetricTimeAppDiscoveryService.java
@@ -18,11 +18,11 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.lametrictime.LaMetricTimeBindingConstants;
-import org.openhab.binding.lametrictime.config.LaMetricTimeAppConfiguration;
-import org.openhab.binding.lametrictime.handler.LaMetricTimeAppHandler;
-import org.openhab.binding.lametrictime.handler.LaMetricTimeHandler;
+import org.openhab.binding.lametrictime.internal.LaMetricTimeBindingConstants;
 import org.openhab.binding.lametrictime.internal.LaMetricTimeUtil;
+import org.openhab.binding.lametrictime.internal.config.LaMetricTimeAppConfiguration;
+import org.openhab.binding.lametrictime.internal.handler.LaMetricTimeAppHandler;
+import org.openhab.binding.lametrictime.internal.handler.LaMetricTimeHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.syphr.lametrictime.api.local.model.Application;
