+++ addons/binding/org.openhab.binding.lametrictime/src/main/java/org/openhab/binding/lametrictime/internal/handler/CountdownAppHandler.java
@@ -6,9 +6,9 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.lametrictime.handler;
+package org.openhab.binding.lametrictime.internal.handler;
 
-import static org.openhab.binding.lametrictime.LaMetricTimeBindingConstants.CHANNEL_APP_COMMAND;
+import static org.openhab.binding.lametrictime.internal.LaMetricTimeBindingConstants.*;
 
 import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
@@ -16,27 +16,27 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.lametrictime.config.LaMetricTimeAppConfiguration;
+import org.openhab.binding.lametrictime.internal.config.LaMetricTimeAppConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.syphr.lametrictime.api.local.ApplicationActionException;
 import org.syphr.lametrictime.api.model.CoreApps;
 
 /**
- * The {@link StopwatchAppHandler} represents an instance of the built-in stopwatch app.
+ * The {@link CountdownAppHandler} represents an instance of the built-in countdown app.
  *
  * @author Gregory Moyer - Initial contribution
  */
-public class StopwatchAppHandler extends AbstractLaMetricTimeAppHandler {
-    private static final String PACKAGE_NAME = "com.lametric.stopwatch";
+public class CountdownAppHandler extends AbstractLaMetricTimeAppHandler {
+    private static final String PACKAGE_NAME = "com.lametric.countdown";
 
     public static final String COMMAND_PAUSE = "pause";
     public static final String COMMAND_RESET = "reset";
     public static final String COMMAND_START = "start";
 
-    private final Logger logger = LoggerFactory.getLogger(StopwatchAppHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(CountdownAppHandler.class);
 
-    public StopwatchAppHandler(Thing thing) {
+    public CountdownAppHandler(Thing thing) {
         super(thing);
     }
 
@@ -44,10 +44,16 @@ public class StopwatchAppHandler extends AbstractLaMetricTimeAppHandler {
     public void handleAppCommand(ChannelUID channelUID, Command command) {
         try {
             switch (channelUID.getId()) {
+                case CHANNEL_APP_DURATION: {
+                    getDevice().doAction(getWidget(),
+                            CoreApps.countdown().configure(((Number) command).intValue(), false));
+                    updateActiveAppOnDevice();
+                    break;
+                }
                 case CHANNEL_APP_COMMAND:
                     handleCommandChannel(command);
-                    updateState(channelUID, new StringType()); // clear state
                     updateActiveAppOnDevice();
+                    updateState(channelUID, new StringType()); // clear state
                     break;
                 default:
                     logger.debug("Channel '{}' not supported", channelUID);
@@ -69,16 +75,16 @@ public class StopwatchAppHandler extends AbstractLaMetricTimeAppHandler {
         String commandStr = command.toFullString();
         switch (commandStr) {
             case COMMAND_PAUSE:
-                getDevice().doAction(getWidget(), CoreApps.stopwatch().pause());
+                getDevice().doAction(getWidget(), CoreApps.countdown().pause());
                 break;
             case COMMAND_RESET:
-                getDevice().doAction(getWidget(), CoreApps.stopwatch().reset());
+                getDevice().doAction(getWidget(), CoreApps.countdown().reset());
                 break;
             case COMMAND_START:
-                getDevice().doAction(getWidget(), CoreApps.stopwatch().start());
+                getDevice().doAction(getWidget(), CoreApps.countdown().start());
                 break;
             default:
-                logger.debug("Stopwatch app command '{}' not supported", commandStr);
+                logger.debug("Countdown app command '{}' not supported", commandStr);
                 break;
         }
     }
