+++ addons/binding/org.openhab.binding.lametrictime/src/main/java/org/openhab/binding/lametrictime/internal/handler/LaMetricTimeAppHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.lametrictime.handler;
+package org.openhab.binding.lametrictime.internal.handler;
 
 import org.syphr.lametrictime.api.local.model.Widget;
