+++ addons/binding/org.openhab.binding.lametrictime/src/main/java/org/openhab/binding/lametrictime/internal/handler/StopwatchAppHandler.java
@@ -6,9 +6,9 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.lametrictime.handler;
+package org.openhab.binding.lametrictime.internal.handler;
 
-import static org.openhab.binding.lametrictime.LaMetricTimeBindingConstants.*;
+import static org.openhab.binding.lametrictime.internal.LaMetricTimeBindingConstants.CHANNEL_APP_COMMAND;
 
 import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
@@ -16,27 +16,27 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.lametrictime.config.LaMetricTimeAppConfiguration;
+import org.openhab.binding.lametrictime.internal.config.LaMetricTimeAppConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.syphr.lametrictime.api.local.ApplicationActionException;
 import org.syphr.lametrictime.api.model.CoreApps;
 
 /**
- * The {@link CountdownAppHandler} represents an instance of the built-in countdown app.
+ * The {@link StopwatchAppHandler} represents an instance of the built-in stopwatch app.
  *
  * @author Gregory Moyer - Initial contribution
  */
-public class CountdownAppHandler extends AbstractLaMetricTimeAppHandler {
-    private static final String PACKAGE_NAME = "com.lametric.countdown";
+public class StopwatchAppHandler extends AbstractLaMetricTimeAppHandler {
+    private static final String PACKAGE_NAME = "com.lametric.stopwatch";
 
     public static final String COMMAND_PAUSE = "pause";
     public static final String COMMAND_RESET = "reset";
     public static final String COMMAND_START = "start";
 
-    private final Logger logger = LoggerFactory.getLogger(CountdownAppHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(StopwatchAppHandler.class);
 
-    public CountdownAppHandler(Thing thing) {
+    public StopwatchAppHandler(Thing thing) {
         super(thing);
     }
 
@@ -44,16 +44,10 @@ public class CountdownAppHandler extends AbstractLaMetricTimeAppHandler {
     public void handleAppCommand(ChannelUID channelUID, Command command) {
         try {
             switch (channelUID.getId()) {
-                case CHANNEL_APP_DURATION: {
-                    getDevice().doAction(getWidget(),
-                            CoreApps.countdown().configure(((Number) command).intValue(), false));
-                    updateActiveAppOnDevice();
-                    break;
-                }
                 case CHANNEL_APP_COMMAND:
                     handleCommandChannel(command);
-                    updateActiveAppOnDevice();
                     updateState(channelUID, new StringType()); // clear state
+                    updateActiveAppOnDevice();
                     break;
                 default:
                     logger.debug("Channel '{}' not supported", channelUID);
@@ -75,16 +69,16 @@ public class CountdownAppHandler extends AbstractLaMetricTimeAppHandler {
         String commandStr = command.toFullString();
         switch (commandStr) {
             case COMMAND_PAUSE:
-                getDevice().doAction(getWidget(), CoreApps.countdown().pause());
+                getDevice().doAction(getWidget(), CoreApps.stopwatch().pause());
                 break;
             case COMMAND_RESET:
-                getDevice().doAction(getWidget(), CoreApps.countdown().reset());
+                getDevice().doAction(getWidget(), CoreApps.stopwatch().reset());
                 break;
             case COMMAND_START:
-                getDevice().doAction(getWidget(), CoreApps.countdown().start());
+                getDevice().doAction(getWidget(), CoreApps.stopwatch().start());
                 break;
             default:
-                logger.debug("Countdown app command '{}' not supported", commandStr);
+                logger.debug("Stopwatch app command '{}' not supported", commandStr);
                 break;
         }
     }
