+++ addons/binding/org.openhab.binding.lametrictime/src/main/java/org/openhab/binding/lametrictime/internal/handler/WeatherAppHandler.java
@@ -6,40 +6,35 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.lametrictime.handler;
+package org.openhab.binding.lametrictime.internal.handler;
 
-import static org.openhab.binding.lametrictime.LaMetricTimeBindingConstants.*;
+import static org.openhab.binding.lametrictime.internal.LaMetricTimeBindingConstants.CHANNEL_APP_COMMAND;
 
-import java.time.Instant;
-import java.time.LocalTime;
-import java.time.ZoneId;
-
-import org.eclipse.smarthome.core.library.types.DateTimeType;
 import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
 import org.eclipse.smarthome.core.types.Command;
-import org.openhab.binding.lametrictime.config.LaMetricTimeAppConfiguration;
+import org.openhab.binding.lametrictime.internal.config.LaMetricTimeAppConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.syphr.lametrictime.api.local.ApplicationActionException;
 import org.syphr.lametrictime.api.model.CoreApps;
 
 /**
- * The {@link ClockAppHandler} represents an instance of the built-in clock app.
+ * The {@link WeatherAppHandler} represents an instance of the built-in weather app.
  *
  * @author Gregory Moyer - Initial contribution
  */
-public class ClockAppHandler extends AbstractLaMetricTimeAppHandler {
-    private static final String PACKAGE_NAME = "com.lametric.clock";
+public class WeatherAppHandler extends AbstractLaMetricTimeAppHandler {
+    private static final String PACKAGE_NAME = "com.lametric.weather";
 
-    public static final String COMMAND_DISABLE_ALARM = "disableAlarm";
+    public static final String COMMAND_FORECAST = "forecast";
 
-    private final Logger logger = LoggerFactory.getLogger(ClockAppHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(WeatherAppHandler.class);
 
-    public ClockAppHandler(Thing thing) {
+    public WeatherAppHandler(Thing thing) {
         super(thing);
     }
 
@@ -47,17 +42,10 @@ public class ClockAppHandler extends AbstractLaMetricTimeAppHandler {
     public void handleAppCommand(ChannelUID channelUID, Command command) {
         try {
             switch (channelUID.getId()) {
-                case CHANNEL_APP_SET_ALARM: {
-                    LocalTime time = Instant.ofEpochMilli(((DateTimeType) command).getCalendar().getTimeInMillis())
-                            .atZone(ZoneId.systemDefault()).toLocalTime();
-                    getDevice().doAction(getWidget(), CoreApps.clock().setAlarm(true, time, null));
-                    updateActiveAppOnDevice();
-                    break;
-                }
                 case CHANNEL_APP_COMMAND:
                     handleCommandChannel(command);
-                    updateState(channelUID, new StringType()); // clear state
                     updateActiveAppOnDevice();
+                    updateState(channelUID, new StringType()); // clear state
                     break;
                 default:
                     logger.debug("Channel '{}' not supported", channelUID);
@@ -78,11 +66,11 @@ public class ClockAppHandler extends AbstractLaMetricTimeAppHandler {
     private void handleCommandChannel(Command command) throws ApplicationActionException {
         String commandStr = command.toFullString();
         switch (commandStr) {
-            case COMMAND_DISABLE_ALARM:
-                getDevice().doAction(getWidget(), CoreApps.clock().stopAlarm());
+            case COMMAND_FORECAST:
+                getDevice().doAction(getWidget(), CoreApps.weather().forecast());
                 break;
             default:
-                logger.debug("Clock app command '{}' not supported", commandStr);
+                logger.debug("Weather app command '{}' not supported", commandStr);
                 break;
         }
     }
